Minor edit & decided to forget image centering.
I probably could pull it off using some remarks from here:
http://stackoverflow.com/a/12118349
but I don't think it looks that bad without the extra work.
I think there's an advantage to keeping the markdown itself simple,
in that it's more pleasant to read directly and that it's more likely to
work well when rendered by different markdown systems.